8/14 update - deletion of categories now moves all associated items to uncategorized category - deletion of "uncategorized" category is now disabled - item models are now in a proper database relationship with events - event page now properly displays items sorted by category and alphabetically - items can now be edited by admins - the username of who created the item is now displayed - ordering whole case functionality mostly complete

+++ groupbuyorganizer/admin/routes.py
@@ -7,6 +7,7 @@ from groupbuyorganizer import database
 from groupbuyorganizer.admin.forms import ApplicationSettingsForm, CreateCategoryForm
 from groupbuyorganizer.admin.models import Category, Instance, User
 from groupbuyorganizer.admin.utilities import admin_check, admin_protector
+from groupbuyorganizer.events.models import Event
 
 
 admin = Blueprint('admin', __name__)
@@ -47,9 +48,19 @@ def category_edit(category_id):
 def category_remove(category_id):
     admin_check(current_user)
     category = Category.query.get_or_404(category_id)
+    if category.id == 1:
+        flash('Access denied', 'danger')
+        return redirect(url_for('admin.category_settings'))
+
+    flash_message = 'Category deleted! '
+    if category.items:
+        flash_message += " Item(s) in this category have been moved to 'Uncategorized.'"
+        for item in category.items:
+            item.category_id = 1
+            database.session.commit()
     database.session.delete(category)
     database.session.commit()
-    flash('Category deleted!', 'info')
+    flash(f'{flash_message}', 'info')
     return redirect(url_for('admin.category_settings'))
 
 
@@ -121,15 +132,33 @@ def app_settings():
     admin_check(current_user)
     instance = Instance.query.first()
     form = ApplicationSettingsForm()
-    #form.registration_enabled.data = instance.registration_enabled
     if form.validate_on_submit():
-        print(instance.registration_enabled)
-        print(form.registration_enabled.data)
         instance.registration_enabled = form.registration_enabled.data
-        print(instance.registration_enabled)
         database.session.commit()
         flash('Changes saved!', 'info')
         return redirect(url_for('admin.app_settings'))
     elif request.method == 'GET':
         form.registration_enabled.data = instance.registration_enabled
     return render_template('app_settings.html', title='Application Settings', form=form)
+
+#todo - temporary for dev purposes
+@admin.cli.command('bootstrap')
+@admin.route("/bootstrap", methods=['GET', 'POST'])
+def bootstrap():
+    database.drop_all()
+    database.create_all()
+
+    database.session.add(Instance())
+
+    database.session.add(Category(name='Uncategorized'))
+    database.session.add(Category(name='Dog Toys'))
+    database.session.add(Category(name='Cat Toys'))
+    database.session.add(Category(name='Fish Supplies'))
+    database.session.add(Category(name='Dog Food'))
+
+    database.session.add(Event(name='Test Session 1'))
+    database.session.add(Event(name='Test Session 2'))
+    database.session.commit()
+
+    flash('Success!', 'primary')
+    return redirect(url_for('general.home'))
